+++ test_autofit/tools/edenise/test_examples/test_unpack_imports.py
@@ -1,7 +1,7 @@
 import pytest
 
 from autofit.tools.edenise import File, Import
-from autofit.tools.edenise.structure import LineItem
+from autofit.tools.edenise import LineItem
 
 
 @pytest.fixture(
