+++ test_autofit/tools/edenise/test_is_in_project.py
@@ -1,4 +1,4 @@
-from autofit.tools.edenise.structure import LineItem
+from autofit.tools.edenise import LineItem
 import pytest
